Banner Service Refactor
After reconsidering the current implementation of the Banner Component, its clear that a refactor to a service is necessary. This illustrates one of the major challenges in changing your mental map from React to Angular. You can't really 'prop drill' in Angular via the @Input() component.

Because you must pass a property from the parent component down to the child via @Input(), creating dynamic content through that decorator feels brittle and limits the reusability of a component. In this case, creating a 'bannerText' propert in the Contacts class definition, then passing it to the Banner component via @Input() feels like it muddies the seperation of logic between components.

Instead of passing data via Import(), my plan is to refactor that logic into a Banner service. Furthermore, I propose using a 'service-in-service' design and further isolating responsibility for Icon selection and rendering to an Icon service. The data for generating the banner text and icon selection will come from a URL slug.

KEYWORDS:
refactor
service
service-in-service
@Import()